# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/imports/{type}/preview": {
      "post": {
        "operationId": "ImportsController_preview",
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "description": "Session-bound CSRF token",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "VALIDATION_ERROR",
                  "message": "Request validation failed",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "UNAUTHENTICATED",
                  "message": "Authentication is required",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "NOT_FOUND",
                  "message": "Resource not found",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "409": {
            "description": "The resource changed; reload and retry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "VERSION_CONFLICT",
                  "message": "The resource changed; reload and retry",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "422": {
            "description": "Manual classification is required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "MANUAL_CLASSIFICATION_REQUIRED",
                  "message": "Manual classification is required",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests; try again later",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "RATE_LIMITED",
                  "message": "Too many requests; try again later",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "provisioning"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MutationRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/imports/{id}": {
      "get": {
        "operationId": "ImportsController_detail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "VALIDATION_ERROR",
                  "message": "Request validation failed",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "UNAUTHENTICATED",
                  "message": "Authentication is required",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "NOT_FOUND",
                  "message": "Resource not found",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "409": {
            "description": "The resource changed; reload and retry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "VERSION_CONFLICT",
                  "message": "The resource changed; reload and retry",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "422": {
            "description": "Manual classification is required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "MANUAL_CLASSIFICATION_REQUIRED",
                  "message": "Manual classification is required",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests; try again later",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "RATE_LIMITED",
                  "message": "Too many requests; try again later",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "provisioning"
        ]
      }
    },
    "/api/v1/admin/imports/{id}/confirm": {
      "post": {
        "operationId": "ImportsController_confirm",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "description": "Session-bound CSRF token",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "VALIDATION_ERROR",
                  "message": "Request validation failed",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "UNAUTHENTICATED",
                  "message": "Authentication is required",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "NOT_FOUND",
                  "message": "Resource not found",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "409": {
            "description": "The resource changed; reload and retry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "VERSION_CONFLICT",
                  "message": "The resource changed; reload and retry",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "422": {
            "description": "Manual classification is required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "MANUAL_CLASSIFICATION_REQUIRED",
                  "message": "Manual classification is required",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests; try again later",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "RATE_LIMITED",
                  "message": "Too many requests; try again later",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "provisioning"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MutationRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "info": {
    "title": "CARE API",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "care_session"
      }
    },
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "code",
          "message",
          "errors",
          "correlationId"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          },
          "correlationId": {
            "type": "string"
          },
          "meta": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "FieldError": {
        "type": "object",
        "required": [
          "field",
          "code",
          "message"
        ],
        "properties": {
          "field": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "MutationRequest": {
        "type": "object",
        "description": "Endpoint-specific request. Aggregate mutations require expectedVersion.",
        "properties": {
          "expectedVersion": {
            "type": "integer",
            "minimum": 1
          }
        },
        "additionalProperties": true
      },
      "SuccessResponse": {
        "type": "object",
        "additionalProperties": true
      },
...
    }
  }
}